Move model_builder directly inside pymc-marketing
