+++ .pre-commit-config.yaml
@@ -1,5 +1,3 @@
-exclude: ^pymc_marketing/vendored/.*$
-
 repos:
   - repo: https://github.com/psf/black
     rev: 22.12.0
