+++ codecov.yml
@@ -14,5 +14,3 @@ coverage:
         if_ci_failed: error #success, failure, error, ignore
         informational: false
         only_pulls: false
-ignore:
-  - "pymc_marketing/vendored/**"
