+++ mypy.ini
@@ -17,9 +17,3 @@ ignore_missing_imports = True
 
 [mypy-scipy.*]
 ignore_missing_imports = True
-
-[mypy-pymc_experimental.*]
-ignore_missing_imports = True
-
-[mypy-pymc_marketing.vendored.*]
-follow_imports = skip
